+++ CMakeLists.txt
@@ -82,8 +82,9 @@ endfunction()
 compiled_test(test/adaptive_storage_test.cpp)
 compiled_test(test/array_storage_test.cpp)
 compiled_test(test/axis_test.cpp)
+compiled_test(test/axis_regular_test.cpp)
+compiled_test(test/axis_variant_test.cpp)
 compiled_test(test/detail_test.cpp)
-compiled_test(test/histogram_dynamic_add_fail.cpp)
 compiled_test(test/histogram_dynamic_fill_one_dimensional_vector_fail.cpp)
 compiled_test(test/histogram_dynamic_fill_one_dimensional_tuple_fail.cpp)
 compiled_test(test/histogram_dynamic_at_tuple_wrong_dimension_fail.cpp)
@@ -91,7 +92,6 @@ compiled_test(test/histogram_dynamic_at_vector_wrong_dimension_fail.cpp)
 compiled_test(test/histogram_dynamic_at_wrong_dimension_fail.cpp)
 compiled_test(test/histogram_dynamic_reduce_wrong_order_fail.cpp)
 # test fail to compile (test/histogram_static_fill_one_dimensional_tuple_fail.cpp)
-compiled_test(test/histogram_static_add_fail.cpp)
 compiled_test(test/histogram_static_at_vector_wrong_dimension_fail.cpp)
 compiled_test(test/histogram_dynamic_test.cpp)
 compiled_test(test/histogram_mixed_test.cpp)
